requirements.txt
flask>=2.3.0
python-dotenv>=1.0.0
openai>=1.0.0
bibtexparser>=1.4.0
